Editor: Move Avatar and AvatarGroup from components to collaborators-presence
The components team wants more time to review before adding new
components to the package. Since Avatar and AvatarGroup are only
consumed by the real-time collaboration UI, move them into the
editor's collaborators-presence folder where they can iterate
independently. They can be promoted back to @wordpress/components
when the team is ready.
