+++ packages/components/src/style.scss
@@ -1,8 +1,6 @@
 @use "@wordpress/base-styles/mixins" as *;
 @use "./animate/style.scss" as *;
 @use "./autocomplete/style.scss" as *;
-@use "./avatar/styles.scss" as *;
-@use "./avatar-group/styles.scss" as *;
 @use "./badge/styles.scss" as *;
 @use "./button-group/style.scss" as *;
 @use "./button/style.scss" as *;
