+++ packages/editor/src/components/collaborators-presence/avatar-group/styles.scss
