+++ packages/editor/src/components/collaborators-presence/avatar/index.ts
