+++ packages/editor/src/components/collaborators-presence/avatar/styles.scss
@@ -1,6 +1,9 @@
 @use "@wordpress/base-styles/colors" as *;
 @use "@wordpress/base-styles/variables" as *;
-@use "../utils/theme-variables" as *;
+
+// Accent color for the avatar. Uses the admin theme color with a hardcoded
+// fallback matching the default WordPress admin blue.
+$-accent-color: var(--wp-admin-theme-color, #3858e9);
 
 .components-avatar {
 	position: relative;
@@ -19,7 +22,7 @@
 	height: $button-size-compact;
 	border-radius: $radius-full;
 	border: 0;
-	background-color: $components-color-accent;
+	background-color: $-accent-color;
 	overflow: clip;
 	flex-shrink: 0;
 	font-size: 0;
@@ -80,7 +83,7 @@
 	grid-template-columns: min-content 0fr;
 	column-gap: 0;
 	padding-inline-end: 0;
-	background-color: $components-color-accent;
+	background-color: $-accent-color;
 	transition:
 		grid-template-columns 0.3s cubic-bezier(0.15, 0, 0.15, 1),
 		column-gap 0.3s cubic-bezier(0.15, 0, 0.15, 1),
