+++ packages/editor/src/components/collaborators-presence/avatar/types.ts
@@ -1,4 +1,4 @@
-import type { IconType } from '../icon';
+import type { IconType } from '@wordpress/components';
 
 export type AvatarProps = {
 	/**
