+++ packages/editor/src/components/collaborators-presence/list.tsx
@@ -1,19 +1,13 @@
 import { __ } from '@wordpress/i18n';
-import {
-	Popover,
-	Button,
-	privateApis as componentsPrivateApis,
-} from '@wordpress/components';
+import { Popover, Button } from '@wordpress/components';
 import { closeSmall } from '@wordpress/icons';
 import { type PostEditorAwarenessState } from '@wordpress/core-data';
 
-import { unlock } from '../../lock-unlock';
+import Avatar from './avatar';
 import { getAvatarUrl } from '../collaborators-overlay/get-avatar-url';
 import { getAvatarBorderColor } from '../collaborators-overlay/avatar-colors';
 
 import './styles/collaborators-list.scss';
-
-const { Avatar } = unlock( componentsPrivateApis );
 
 interface CollaboratorsListProps {
 	activeCollaborators: PostEditorAwarenessState[];
