+++ packages/editor/src/style.scss
@@ -6,6 +6,8 @@
 @use "@wordpress/media-editor/build-style/style.css" as *;
 @use "./components/autocompleters/style.scss" as *;
 @use "./components/collab-sidebar/style.scss" as *;
+@use "./components/collaborators-presence/avatar/styles.scss" as *;
+@use "./components/collaborators-presence/avatar-group/styles.scss" as *;
 @use "./components/collapsible-block-toolbar/style.scss" as *;
 @use "./components/block-visibility/style.scss" as *;
 @use "./components/blog-title/style.scss" as *;
